+++ src/routes/index.tsx
@@ -1,6 +1,10 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { KelsierPage } from "../components/kelsier/KelsierPage";
 import { getActiveAssessmentQuestionnaire } from "../server/assessmentQuestionnaire.functions";
+import {
+	deleteGuestAttempt,
+	startGuestAssessment,
+} from "../server/guestAssessment.functions";
 
 export const Route = createFileRoute("/")({
 	loader: () => getActiveAssessmentQuestionnaire(),
@@ -22,5 +26,14 @@ export const Route = createFileRoute("/")({
 function HomeRoute() {
 	const questionnaire = Route.useLoaderData();
 
-	return <KelsierPage questionnaire={questionnaire} />;
+	return (
+		<KelsierPage
+			questionnaire={questionnaire}
+			persistenceActions={{
+				startAttempt: () => startGuestAssessment(),
+				deleteAttempt: (attemptId) =>
+					deleteGuestAttempt({ data: { attemptId } }),
+			}}
+		/>
+	);
 }